Rename component for consistency.
